requirements.txt
trendspy>=0.0.5
pandas>=1.3.0
schedule>=1.1.0
backoff>=2.1.0
python-dotenv>=0.19.0
urllib3<2.0.0  # 使用1.x版本避免SSL警告